// src/Mobile/ReportsApp/services/api/models/ResumenReportConfiguration.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Inspections API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ResumenReportConfiguration
 */
export interface ResumenReportConfiguration {
    /**
     * 
     * @type {number}
     * @memberof ResumenReportConfiguration
     */
    id?: number;
    /**
     * 
     * @type {number}
     * @memberof ResumenReportConfiguration
     */
    type?: number;
    /**
     * 
     * @type {string}
     * @memberof ResumenReportConfiguration
     */
    title?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ResumenReportConfiguration
     */
    formName?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ResumenReportConfiguration
     */
    remarksLabelText?: string | null;
    /**
     * 
     * @type {number}
     * @memberof ResumenReportConfiguration
     */
    definedCheckLists?: number;
    /**
     * 
     * @type {number}
     * @memberof ResumenReportConfiguration
     */
    definedSignatures?: number;
    /**
     * 
     * @type {number}
     * @memberof ResumenReportConfiguration
     */
    usedByReports?: number;
    /**
     * 
     * @type {boolean}
     * @memberof ResumenReportConfiguration
     */
    inactive?: boolean;
    /**
     * 
     * @type {string}
     * @memberof ResumenReportConfiguration
     */
    printSection?: string | null;
}

export function ResumenReportConfigurationFromJSON(json: any): ResumenReportConfiguration {
    return ResumenReportConfigurationFromJSONTyped(json, false);
}

export function ResumenReportConfigurationFromJSONTyped(json: any, ignoreDiscriminator: boolean): ResumenReportConfiguration {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': !exists(json, 'id') ? undefined : json['id'],
        'type': !exists(json, 'type') ? undefined : json['type'],
        'title': !exists(json, 'title') ? undefined : json['title'],
        'formName': !exists(json, 'formName') ? undefined : json['formName'],
        'remarksLabelText': !exists(json, 'remarksLabelText') ? undefined : json['remarksLabelText'],
        'definedCheckLists': !exists(json, 'definedCheckLists') ? undefined : json['definedCheckLists'],
        'definedSignatures': !exists(json, 'definedSignatures') ? undefined : json['definedSignatures'],
        'usedByReports': !exists(json, 'usedByReports') ? undefined : json['usedByReports'],
        'inactive': !exists(json, 'inactive') ? undefined : json['inactive'],
        'printSection': !exists(json, 'printSection') ? undefined : json['printSection'],
    };
}

export function ResumenReportConfigurationToJSON(value?: ResumenReportConfiguration | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'type': value.type,
        'title': value.title,
        'formName': value.formName,
        'remarksLabelText': value.remarksLabelText,
        'definedCheckLists': value.definedCheckLists,
        'definedSignatures': value.definedSignatures,
        'usedByReports': value.usedByReports,
        'inactive': value.inactive,
        'printSection': value.printSection,
    };
}